Further fix for shader linkage on Linux/OpenGL
The earlier fix (which remains necessary on AMD+Mesa, although
apparently not on nVidia) moved some required shader function
definitions into _GetOSDCommonShaderSource(), however it failed to move
certain forward decls (for GetProjectionMatrix() and GetTessLevel())
which are needed by the moved functions. This resulted in compilation
issues once OsdProjectionMatrix() and OsdTessLevel() tried to call
functions which had not been forward declared (and whose definitions
appear later in the shader source).